Remove Ops traits, and just add F to methods, per review
One troubling point of making this change, is the breaking of type
inference. Because no IO is passed to the create methods, the IO
has to be specified in the type arguments, which means all types
must be specified. So you get this troubling issue as below, where you
have to recite the argument type:

    Handler.create[IO, Option[VDomModifier]](Option(div("a"))).flatMap { myOption =>

This can be fixed using the partial type application trick, so one can
just write
    Handler.create[IO](Option(div("a"))).flatMap { myOption =>

See {Handler, Store}.CreatePartiallyApplied