fix conversion for VMod.attr types
In snabbdom, patching an attribute depends on the type.
If is a boolean, the will add attribute as empty string (for true) and remove attribute (for false).
All others will be handed to the element.setAttribute api, which does `toString` on anything that is not a string.

So, we can either give a Boolean or a String to snabbdom.

What this fixes?
Do not convert Boolean to String for all `VMod.attr[Boolean](name)` -> Instead pass the boolean to snabbdom.